rename minecraft module package paper -> minecraft
