+++ minecraft/src/main/java/love/broccolai/corn/minecraft/item/ItemBuilder.java
@@ -1,4 +1,4 @@
-package love.broccolai.corn.paper.item;
+package love.broccolai.corn.minecraft.item;
 
 
 import org.bukkit.Bukkit;
