+++ minecraft/src/main/java/love/broccolai/corn/minecraft/item/package-info.java
@@ -1,4 +1,4 @@
 /**
  * Utilities for manipulating items
  */
-package love.broccolai.corn.paper.item;
+package love.broccolai.corn.minecraft.item;
