+++ minecraft/src/main/java/love/broccolai/corn/minecraft/item/special/package-info.java
@@ -1,4 +1,4 @@
 /**
  * Builders specific to certain {@link org.bukkit.inventory.meta.ItemMeta}s.
  */
-package love.broccolai.corn.paper.item.special;
+package love.broccolai.corn.minecraft.item.special;
