+++ minecraft/src/main/java/love/broccolai/corn/minecraft/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Paper-specific utilities
+ */
+package love.broccolai.corn.minecraft;
