--- minecraft/src/main/java/love/broccolai/corn/paper/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Paper-specific utilities
- */
-package love.broccolai.corn.paper;
